refactor: structure utils at root
