+++ src/cli/index.ts
@@ -1,8 +1,9 @@
 #!/usr/bin/env node
 import * as path from 'path';
-import { ProcessRetriever, executeScript } from "./utils";
+import { executeScript } from "./utils";
+import ProcessRetriever from "./Process";
 import commandsMap from "./constants";
-import { getDirectories } from '../utils';
+import { pathTo } from '../utils';
 
 const cli = new ProcessRetriever(process, commandsMap);
 
@@ -10,7 +11,7 @@ const cli = new ProcessRetriever(process, commandsMap);
 const { action } = commandsMap[cli.command()];
 
 if (action[0]) {
-    executeScript(path.join(getDirectories().package, action[1]));
+    executeScript(path.join(pathTo().package, action[1]));
 } else {
     console.log(action[1]);
 }
